Fixed controllers

+++ lib/app_web/controllers/comment_controller.ex
@@ -3,6 +3,7 @@ defmodule AppWeb.CommentController do
 
   alias App.Thread
   alias App.Thread.Comment
+  alias AppWeb.Utils.ErrorUtils
 
   action_fallback AppWeb.FallbackController
 
@@ -13,11 +14,11 @@ defmodule AppWeb.CommentController do
     render(conn, "index_with_comment.json", comments: comments)
   end
 
-  def create(conn, %{"comment" => comment_params}) do
-    with {:ok, %Comment{} = comment} <- Thread.create_comment(comment_params, conn.assigns.user, conn.path_params["post_id"]) do
+  def create(conn, %{"post_id" => post_id, "comment" => comment_params}) do
+    with {:ok, %Comment{} = comment} <- Thread.create_comment(comment_params, conn.assigns.user, post_id) do
       conn
       |> put_status(:created)
-      |> put_resp_header("location", comment_path(conn, :show, comment))
+      |> put_resp_header("location", post_comment_path(conn, :show, post_id, comment))
       |> render("show.json", comment: comment)
     end
   end
@@ -37,8 +38,14 @@ defmodule AppWeb.CommentController do
 
   def delete(conn, %{"id" => id}) do
     comment = Thread.get_comment!(id)
-    with {:ok, %Comment{}} <- Thread.delete_comment(comment) do
-      send_resp(conn, :no_content, "")
+
+    if conn.assigns.user.id == comment.user_id do
+      with {:ok, %Comment{}} <- Thread.delete_comment(comment) do
+        send_resp(conn, :no_content, "")
+      end
+    else
+      conn
+      |> ErrorUtils.throw(:unauthorized, reason: "Cannot delete other user comment.")
     end
   end
 end
